+++ apps/docs/src/VERSIONS.ts
@@ -11,6 +11,10 @@ export const CORE_VERSIONS = [
   "0.7.0",
   "0.7.1",
   "0.7.2",
+  "0.7.3",
+  "0.7.4",
+  "0.8.0",
+  "0.8.1",
 ].reverse();
 
 export const LATEST_CORE_CHANGELOG_URL = `/docs/changelog/${CORE_VERSIONS[0].replaceAll(".", "-")}`;